Stop generating all prover call at once
Instead of generating all prover calls (which can be lengthy on setups
with a lot of processes used by default) in a timeout, we generate
only one in an idle call (which is called more often).
The expected result is to avoid freezing of the IDE when the user has a
lot of processors.

Use prio of 300 for the idle call: from testing, using less than 100 take
priority over GTK (resulting in freezing).

Renaming timeout_handler to idle_handler