S708-024 Run external prover without step limit (limit_steps = 0)
When a steps limit is given for a prover (which is always in SPARK ?), the
part of the configuration that is used to generate the command is
"command_steps" (see get_complete_command in src/driver/whyconf.ml). This
command is not defined for any potential external prover (see
share/prover-detection-data.conf) so ext-provers are not usable. This
patch sets the steps limit of all non builtin provers to 0 (force to use
"command").

* src/gnat/gnat_config.ml
(convert_data): Remove altergo from the default.
(find_convert): Add comment

* src/gnat/gnat_objectives.ml
(run_goal): If commands_steps does not exist for the prover, steps are
removed. Otherwise, an Highfailure would be returned in
get_complete_command of Whyconf.

Change-Id: I2fea42d721382e87388dd454b5b8890939e3febb