S716-015 Add a more complete set of locations to each variables
Each variables generated by counterexamples may have several locations.
This adds several locations to variables that are found on their duplicate
so that the counterexamples exhibits more values.
This will also add more badly located counterexamples (and counterexamples
on leftvalues). This was accepted by Yannick on discussion under S716-015.

* src/printer/smtv2.ml
(print_info_model): When saving the assignments from terms to the string
used as corresponding smtv2 constant, we do not erase terms that appear
several times. Instead, we copy the locations that appear several times on
the term that is kept.

Change-Id: I53937560fad8c6ec6ca78cc7a9ecfc79d7bff1e5