S327-015 Adding a simpl tactic for SPARK generated task
This adds a new easy to use transformation that should simplify the use of
manual proof by providing a tactic that is a composition of introductions,
simplifications of trivial, trivial let bindings, and classic
simplification of task.

* src/transform/gnat_trivial.ml
(eliminate_trivial_let): Transformation that simplifies all trivial lets
of the form "let A = B in" with B being a single variable/identifier.
(spark_simpl): Easy to use simplification transformation


Change-Id: I1eb32a58090a6b78770f69a5bc2803c9ede00b0f