S821-001 Printer errors should be reported as warnings
This reports printer/internal Why3 error during elaboration of a prover
file as a warning.
Note that seeing one of these warning may break the way the gnatwhy3
scheduling of prover works. In particular, such an error was the cause of
the absence of the execution of counterexamples provers.

* src/gnat/gnat_main.ml
(interpret_result): Add a case for internal error in the matching of the
callback used by gnatwhy3.

Change-Id: Ie30f5b3bcc8b770402edc915dcdbd54b8845d7ed