S911-045 Patch the way debug flag are used
This creates a new formatter which register all printed information in
a buffer defined in gnatwhy3. The formatter is used for all Debug related
functions now. So, all Why3 debug information should now be printed as
warnings at the end of the run of gnatwhy3 (this may be a lot of memory
in some case but we don't care because it is a debug flag that is not
supposed to be used by end-users).

* src/gnat/gnat_main.ml
(escape_buffer): Buffer containing all that is printed to escape_formatter
(escape_formatter): formatter associated to above buffer. The
set_debug_formatter part make the formatter used by default in the Why3
debug system.
(ending): Print the contents of the Buffer as a warning.

Change-Id: Ib98a865e58812d975bf7b98f9262ca0dadbbe501