external_printer: Add a way to change the keywords highlights of task
This is done dynamically: when sending a task, the server also send the
name of the language it was done in. The ide changes the language used by
the task_view on each task.

This also adds an XML conf for keywords of task originating from python
(why3py.lang).