error location: add wrapper for import_scope/namespace
The wrapper are used to add calls to Loc.try on import_scope/ns.
This improves errors "unknown scope" which are now more precise and
target the qualid instead of the whole encapsulated function.

Also fix Loc.join:
when b1 < a2, (Loc.join [a1; b1] [a2; b2]) = [a1; b2]
When not on the same line a heuristic is used