SB26-016 Activate Ada external printing/parsing
This activates the external printer/parser so that task are printed as
Ada terms and they can be parsed the same in arguments of transformations.

Note that this feature can be disactivated by uncommenting the code in
args_wrapper and commenting the additional registration in ada_main.ml

* Makefile.in
Make the ada plugins a new dependency of gnatwhy3/gnat_server.

* plugins/ada_terms/ada_main.ml
Activate parsing/printing through registered function.

* src/transform/arg_wrapper.ml
Disactivate the registering of the standard whyml_format parsing.

Change-Id: I295a7074c74e606875da31f514253f5f49f355aa