transformation: replace now understands coercions
This is now allowed to make transformations that take a list of
arguments of same type (and args_wrapper ensures coercions are applied so
that the elements are of the same type).

This is done with a *hack* in args_wrapper.ml: for Ttermlist_same, a list
of elements [a1; a2; a3; ...; an] is typed by creating a bigger elements
T = (a1 = a2 /\ a1 = a3 /\ ... a1 = an). Typing T using standard typing
creates correctly coerced terms.