Test Why3 without Opam.
In other words, only native Debian packages are used for the compilation.

This commit also avoids compiling Alt-Ergo several times by installing it
globally. It also moves the Re dependency so that it is tested only by
the "full" target. It also upgrades the compiler to 4.08.1 for "full".
Finally, it also tries to build why3-coq during the "opam" target.