Use incremental maps to store identifiers when computing checksums.
Instead of memoizing the whole map of identifiers for each task
declaration, only the difference with the parent task or with one of the
child tasks is stored. The most recently accessed map, however, is fully
stored. When generating checksums, the number of differences that have to
be taken into account to compute the full map is amortized O(1).

On multiprecision/toom, the memory consumption is reduced by 15% while the
running time is unchanged.