PA07-033 Print and report type errors in generated ptree by markers
The locations in the generated ptree are unique but useless because there is
no corresponding concrete syntax.

With this change, if a typing error is encountered in the generated ptree,
we instruct the printer for the mlw-file to mark the corresponding node
by (*XXX*). The type error is reported by filename and a hint to look for
the mark.

Change-Id: I0b8e3a121f3c4f9b82954ebc4c38c3aae6bcfddc