T211-001 add option --list-transforms to gnatwhy3
This option will be used in SPARK build to generate the list of
transformations in a JSON file.
It's important that this happens early in gnatwhy3 and gnatwhy3
shouldn't do anything else, as the binary isn't properly installed
when called, and initializiation would likely fail.

* gnat_config.ml
(output_list_transforms) new function to print the list of
transformations when --list_transforms was given.

Change-Id: I82a052a84466794b8161f0a8c065823e6cb6c828