Avoid unnecessary copying of large binaries
When executing 'make install' we copy ~250MB of OCaml binaries, which takes
a small but noticeable amount of time (and most likely isn't great for the
hard discs). Now binaries are only copied when they differ from the existing
files.

* Makefile.in
(install_spark2014): Replace "cp" with "install -C", which is documented as
"compare each pair of source and destination files, and in some cases, do
not modify the destination at all".

no-tn-check

Change-Id: Ib59521e57e020fed32979250792c2c8d1ce2de0e