Choose representation of arrays as terms by environment variable
An array in the runtime of pinterp is an OCaml array of pinterp values. An array in the
Why3 logic is a private record with fields `length` and `elts`, mapping indices to values.

A runtime array `a` of length `n` can be converted into a logical Why3 array as

1. a sequence of updates: `(make <n> undefined)[0 <- a[0]]...[n-1 <- a[n-1]]`
2. an epsilon term: `(epsilon v. v.length = <n> /\ v[0] = a[0] /\ ... /\ v[n-1] = a[n-1])`

From a quick test it seems that provers can handle arrays much better as epsilon
terms (because it is flatten out by transformation `eliminate_epsilon`.

So by default, arrays are converted into an epsilon term, but if the environment variable
`WHY3PINTERPARRAY` is set to `update`, it is converted in a sequence of updates.