Add a wrapper around Dynlink.
Starting with OCaml 4.08 Dynlink intercepts exceptions raised by dynamic
libraries at initialization time and wraps them. This commit unwraps the
exception, thus recovering its actual content.

Unfortunately, this is not sufficient to get --debug=stack_trace to
produce the correct backtraces, due to a bug in OCaml fixed in 4.11.

In other words, backtraces are broken for OCaml between 4.8 and 4.10.