Fix continuous integration (partly revert 333c915e).
Ocamlfind is not installed on the "system" setup, so we cannot use it to
locate Zarith. In fact, Zarith might not even be installed, which means
that Num should be used instead. So, we use the same mechanism as other
extraction tests.

This commit also restores proper error messages in case of failure.