Add support for Broadway in the deployed image.
The wrapper script checks for the WHY3IDE environment variable to decide
whether it should intercept the "ide" command and launch broadwayd.

Pango, as shipped in the deployed image, does not seem to appreciate
font size specifications, thus causing text to appear only in menus and
dialogs. So, we disable this feature when the GDK_BACKEND environment
variable is set to "broadway".