S906-022 improve search for labels in why3 tree
Instead of looking at one specific node in the Why3 tree, we
traverse the entire "right-hand side" of the tree and collect all
relevant labels. This leads to more robust identification of labels.

Also, when looking for labels in the VC, we also extract the "extra
info" (currently the Ada node used for pretty printing), and store
it in the map of VCs for each check. This avoids traversing the VC
again to extract this info later during reporting.

* gnat_expl.ml
  - introduce new type vc_info that groups the check and the extra
    info
  - searching for labels now yields a vc_info object and not just a
    check
  - search looks at every node from the top, descending into the
    right side of the term.
* gnat_objectives.ml
  - adapt to new record structure
  - change map (goal -> check) to (goal -> vc_info)
* gnat_report.ml
(register): now directly take the extra info as argument, instead of
  the task to search for it
* gnat_main.ml
  - when goal is unproved, query map in gnat_objectives.ml to get
    vc_info and pass the extra info to call to Gnat_report.register
* gnat_server.ml
  - adapt to new record structure

Change-Id: I31383ea53c5feee91225dcef5b46ff1b0718c35c