Use a single Isabelle driver for realizations.
The file lib/isabelle/realizations.xxx was updated depending on the
version of Isabelle detected at configure time. As a consequence, it was
too easy to mistakenly update only one of the files. Yet, both files were
identical since both drivers were the same. So, the added complexity was
not warranted.