Complete missing theory specializations for 128bits bitvectors
The work done on T824-034 to support 128bits integers in GNATprove
did not specialize the new theories for SMTLIBv2. Now done.

* drivers/smt-libv2-bv.gen: Add theory specialization for all
missing conversions from 8/16/32/64 bits from/to 128 bits, as well
as between 128 bits and 256 bits. These correspond to the clones of
theory BVConverter_Gen that were added in bv.mlw. Also add a theory
specialization for the clone of BV_Gen for 256 bits that was also
added to bv.mlw.

Change-Id: Icf163d356dcb86541305e3ea1f3accaccef7fb67