U802-011 fix incorrect sloc for inlined nodes
If a pretty-printing node was obtained through inlining, we now also
return the node of the inlined expression. This can be used for
better placement of VCs for pretty-printing in gnat2why.

* gnat_expl.ml
- "inlined" field is now an int option instead of a boolean.
(read_vc_labels): Set inline field to previous pretty node if
  available, and to -1 (indicating "True") otherwise.
* gnat_main.ml
 adapt default value for extra_info
* gnat_report.ml
(print_info): print node info for inlined node if available

Change-Id: Id40353d9c258cb33ba37fc60533dd8fcd794fabf