Keep trivial VC declarations with vc:annotation anywhere
Commit 3185e9a9 changed the detection of trivial VC declarations with the effect
that trivial VC declarations are discarded when the [@vc:annotation] was not on
the term root but on a descendant term. For example in the code below f2'vc is
dropped, because the term in the declaration of the postcondition VC becomes
`let result = 42 in [@vc:annotation] true`. This behaviour resulted in lost
checks in SPARK.

```
use int.Int

let f1 () : (result: unit)
    ensures { true }
= ()

let f2 () : int
    ensures { true }
= 42
```

To resolve this and retain trivial VC declarations, this commit searches for
vc:annotation in the term deeply.