UB04-014 save temp for new value of variant
When generating the assertion comparing the old value of the variant
with the new one, we were directly using the expression provided in
the Variant node of the GNAT AST. This is not correct, as it may be
a program expression, so not suitable for an assertion. We introduce
let-bindings to introduce temps for these expressions.

* gnat_ast_pretty.ml
  new pretty printing functions to (almost) complete pretty-printing
  of nodes
* gnat_ast_to_ptree.ml
  (mk_of_expr): introduce temps for new values of variants

Change-Id: Ifeb30d508ec834ddf0253c19859f5305fe18d8a5