UC17-044 Remove axioms related to power for Alt-Ergo
The recent reinstalment of these axioms is detrimental to the proof
of code using exponentiation in the runtime.

Experiments show that CVC4 and Z3 are also impacted by these axioms,
but removing them, or even the ones with no reasonable triggers
(Power_monotonic and Power_comm1) that impact proof of some VCs, or
just adding triggers on the remaining axioms, all lead to regressions
on other VCs. No change thus for now for these provers.

Change-Id: I49d3fc1ecf6cf1ff5e82e835114a4f24a89b8a76