V107-011 improve inlining in gnatwhy3
When inserting a pair (function symbol, expression) into the
environment map used for inling, the expression should get the
GP_Inlined flag so that inlining information can be passed on
properly to gnat2why. This information was missing in one case. This
case happens to be dead code right now but it might be used if Why3
code generation changes slightly.

Also improved some comments on the JSON result format, and removed a
duplicate comment.

* gnat_report.mli
comment improvement
* gnat_split_conj.ml
(fold): add label to expression inserted into env

Change-Id: I83f2bc4f1e0bc155e1be8468f1d4093f60bf4936