(no-tn-check) refactor code
The two functions is_most_appropriate_prover and
select_appropriate_proof_attempt shared a lot of logic, and the
latter is the only function to call the former.
is_most_appropriate_prover can be removed and a small part of it is
added to select_appropriate_proof_attempt.

* gnat_objectives.ml
(is_most_appropriate_prover): removed
(select_appropriate_proof_attempt): replaced calls to
  is_most_appropriate_prover by simpler expression

Change-Id: I9462be559a5d0cabd7ed3cb8745a1ec37cb921af