No longer hardcode the version into prover filenames.
The directories from the PATH environment variable are now scanned to find
files that have either the same name as an "exec" field (modulo ".exe" and
".bat" suffixes) or have it as a prefix followed by a dash and a digit.
Also, the fully qualified name of the prover is now stored into why3.conf.