V330-007 add code to remove duplicate includes
IN an effort to clean up the generated mlw, we remove duplicate
includes. We do this by transforming the includes in an intermediate
representation which can be easily be compared, then use a set to
avoid duplicates. We do this in a way as to preserve ordering of the
includes, as why3 is apparently sensitive to that.

* gnat_ast_to_ptree.ml
(Include_Decl): new module to be able to build sets of includes
(mk_theory_declaration): build a set of includes before translation,
  to avoid duplicates

Change-Id: Ibd06f7b7edfe3c55ed5e467837c7f021d17e052e