V502-004 new transformation to remove unused symbols
We introduce a new transformation that simply removes unused
symbols. It is currently enabled only on z3. This is because z3 VCs
initially motivated this transformation, and z3 contains the
"inline_premises" transformation which particularly benefits from
this elimination of unused symbols. We can experiment later with
enabling this transformation for other provers as well.

eliminate_unused.ml:
  new transformation that traverses a task and removes any symbols
  that were not used at all
z3_gnatprove.drv:
  enable transformation

Change-Id: I3fda16cd3eb983e130d33c30e05e15790204924f