Display source files in the IDE
When a source location points to a file that is not part of the session,
the IDE now displays it in an extra tab. If the source locations are
properly positioned, then the source positioning and coloring should
work in these extra files. This feature should be useful for front-ends
generating mlw files with being recorded as a language plug-in