Split Check_ce.select_model to expose more functions in the API
The idea is to be able to call get_rac_results (eventually only for giant-step RAC)
and to provide several functions in the API to select a model: based on the the
verdict, or only on the giant-step RAC results (the latter remains to be implemented).

Tests still have to be updated (thus skipping CI).