Fix usage of xmllint with newer versions.
Irrespective of option --dtdvalid, xmllint now looks up the DTD file on
the network. Passing --nonet prevents this behavior but xmllint then
complains that it is not allowed network access. Using --path and --valid
does not prevent xmllint from complaining about the network, but it at
least avoids a complaint about http://... not being a valid filename.