V429-003 rewrite simple to_array/of_array chains
We add a new transformation to Why3 that simplifies expressions
of the form
  to_array (x)
  first (x)
  last (x)

where x is either "of_array (a, f, l)", or a variable that was
previously introduced via a let-binding, and the definition of the
let-binding is of that form.

We recognize the symbols by special attributes. In this way,
the transformation works across clones of these symbols, but won't
apply to any other to_array/of_array symbols.

* gnat_split_conj.ml
- remove unused code
- add new transformation to the standard transformation for
  gnatprove goals
* gnat_rewrite.ml
  detect applications of the aforementioned symbols and rewrite if
  appropriate. Store let-bindings in a special environment so that
  the rewriting works across let-bindings.

Change-Id: Ie37b9cdd6fd439697076066617239eb95056d9d3
Depends-On: I7e2e5853cb59d01695e7cce511f598ee0d61d908