V914-015 kill processes when why3server interrupted
On Linux, we create new process groups for provers to be able to
kill provers including their potential child processes. As a
consequence, if the process group of why3server is killed, the
provers are *not* killed automatically. We therefore install a
handler to kill prover processes when why3server receives a signal
to terminate.

* proc.c
(kill_all_processes): new function to kill prover processes when
  why3server is interrupted
* server-unix.c
(sigterm_handler): new function to be registered as handler for
  shutdown
(set_sigterm_handler): new function to register the handler
(server_init_listening): call set_sigterm_handler to set up graceful
  shutdown

Change-Id: Ie0e2d486428c9d00f617f998e7bbf500f4a96674