VB30-001 fix unfolding for explanations
The unfolding transformation tries to detect "defining axioms" of
this form:

for all v1 ... vn. f v1 ... vn = def

and bases the computation of the substitution on the quantified
variables (instead of the function arguments) assuming that the
quantified variables are in the same order as the function arguments.
It turns out they are not always in the same order, which resulted in
type mismatches when doing the substitution. It's better to base the
substitution on the function arguments which are guaranteed to be in
the correct orde

* gnat_split_conj.ml
(unfold_right): only unfold a symbol if defined in unfolding map
(is_all_vars): renamed to all_vars: extract the variables of a list
  of terms, if all terms are just variables
(extract_def): remove vs argument; extract variables from list of
  terms using all_vars
(extract_def_from_axiom): don't use the variable list from the
  quantifiers
(unfold_trans): remove catch-all exception handler

Change-Id: I8cfadfc416d1efd87eb4bc11a96243efda05cb6d