(no-tn-check) fix location of log file
why3server changes the directory before accessing the socket file,
and then changes back.  The creation of the log file happens
(unnecessarily) in the window where the directory is changed, which
leads to the log file being created in the same directory as the
socket file. We move the log file creation before the directory
switch.

* server-unix.c
(server_init_listening): move init of log file earlier

Change-Id: If01bf323f5f5d1a77ba2f94cf17a3fbfeb36321b