Eliminate algebraic: support for the algebraic:kept attribute.
This attribute instructs elmiminate_algebraic to keep primitive
operations (pattern matching, constructors and projections) on these
types.

A new option of the transformation, keep_m, instructs it to keep all
the monomorphic algebraic datatypes.

The rest of the chain is adapted to support monomorphic algebraic types.

For now, the only type tagged with this attribute is bool.