W123-017 Fix Not_found exception in gnatwhy3
In replay, the CE goals are not properly registered. They are registered
when created in a normal run, but the replay run finds them already in
the session and the creation code is not run. We robustify the code that
attempts extracting info from these goals, by checking parent goals in
this case.

* gnat_main.ml
(report_messages): protect against the situation with CE goals, by
  attempting to extract info from parent goal
* gnat_objectives.ml
- fix Pervasives -> Stdlib
- export is_ce_goal