don't use default value when reading in proof results
Why3 might store a proof attempt without result in the session file, for
example when Why3 was interrupted before the prover returned. But when
reading such session files, an absence of result was conflated to a
"failure" result. This would make it impossible to distinguish between
an actual proof failure and an unfinished/absent proof.