Remove unused constants in SPARK VCs
We replace the "remove_unused_keep_constants" transformation by
"remove_unused", to also replace unused constants. We can do this as
constants which should not be removed already have a special flag
attached to them.

To be able to change the transformations for cvc5, we need to introduce
a file "cvc5.gen" which contains all cvc5 information *except* the
transformations, so we can include it from cvc5_gnatprove.drv.