Make it easier to support statically linked commands.
Commands are now registered inside a table. If a given command is not
found inside the table, the corresponding plugin is loaded (as before),
and then another lookup is performed inside the table. Once a command is
found, the registered module is executed (as before).

This commit also fixes some missing module for the following commands:
doc, ide, session, wc, web.